Move environment variable handling
Move environment variable out to its own value provider. It is complex relative to the other properties handled by the `ActiveLaunchProfileCommonValueProvider` and none of the code is shared by other properties.
